+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.4.3 2022-02-08
+
+* extract correlation id from message header
+* apply clippy recommendations
+
 # 0.4.2 2021-12-08
 
 * truncate message body buffer for binary messages